refactor: move tags
