+++ packages/components/tag/index.js
@@ -1,3 +1,3 @@
-export { default } from '../../../src/components/tag';
+export { default } from './src/tag';
 
 export { default as version } from './version';
