+++ packages/components/tag/src/tag.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../test-utils';
+import { render } from '../../../../src/test-utils';
 import Tag from './tag';
 
 it('should render children', () => {
