+++ packages/components/tag/src/tag.story.js
@@ -3,9 +3,9 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { withKnobs, boolean, select, text } from '@storybook/addon-knobs/react';
-import Section from '../../../.storybook/decorators/section';
+import Section from '../../../../.storybook/decorators/section';
 import Tag from './tag';
-import Readme from './README.md';
+import Readme from '../README.md';
 
 const Story = () => (
   <Section>
