+++ packages/components/tag/src/tag.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ThemeProvider } from 'emotion-theming';
 import { Tag } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../test/percy';
+import { Suite, Spec } from '../../../../test/percy';
 
 const longText = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur et
 metus ultrices, interdum augue eget, consequat orci. Nam et nisi
